CAMEL-19400: camel-elasticsearch - Use nested classes to speed up
